pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-automation-orchestrator"
version = "0.1.0"
description = "Unified platform combining OpenHands, crewAI, AutoGPT, Strix, and MoneyPrinterV2"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = {text = "MIT"}
dependencies = [
    "crewai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black>=23.0", "ruff>=0.1.0", "mypy>=1.0"]

[project.scripts]
aiauto = "orchestrator.cli.main:cli"

[tool.setuptools.packages.find]
include = ["orchestrator*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
